Convert a TOA reflectance image to a TOA radiance image

Converts a TOA reflectance image to a TOA radiance image.

refl_to_rad(toa_refl, date, esun, solar_zenith_angle)

Arguments

toa_refl
A stars image with top of atmosphere reflectance values

date
Acquisition date, needs to be Date of length 1

esun
External solar radiation [W/m^2]. For Venus satellite bands, esun values are provided in the bands dataset.

solar_zenith_angle
Solar zenith angle [degrees, between -180 and 180]

Value

A stars image with top of atmosphere reflectance values

Details

Top of atmosphere radiance can be caulculated using the following formula:

L = TOA * (ESUN * cos(solar_zenith_angle) / (pi * d^2)

where 'L' is the top of atmosphere radiance, 'TOA' is the top of atmosphere reflectance, 'ESUN' is the external solar radiation and 'solar_zenith_angle' is the solar zenith angle.

Examples

# Band 3 in sample image 'venus1'
library(stars)
library(XML)

# Reflectance Venus image, band 3
refl = venus1[,,,3,drop=TRUE] / 1000

# Metadata
l = xmlToList(venus1m)
m = get_venus_metadata(l, band = 1)

# Specific pixel value: reflectance
refl[[1]][100,100]  

#> [1] 0.133

# Conversion from reflectance to radiance
rad = refl_to_rad(
    toa_refl = refl,
    date = m$date,
    esun = bands$esun[3],
    solar_zenith_angle = m$solar_zenith
)

# Specific pixel value: radiance
rad[[1]][100,100]

#> [1] 48.67207