6S parameters

Retrieves 6S radiative transfer model paremeters from the 6S web interface https://www-loa.univ-lille1.fr/Wsixs/. The parameters can be used to atmospherically correct VENµS satellite images.

sixs_params(
  remote_driver,
  day,
  month,
  SolarZenithalAngle,
  SolarAzimuthalAngle,
  ViewZenithalAngle,
  ViewAzimuthalAngle,
  Longitude,
  Latitude,
  Uw,
  Uo3,
  opticalDepth,
  LowerWavelength,
  UpperWavelength,
  TargetAltitude,
  GroundCondition = c("Homogenous Ground", "Patchy Ground"),
  TargetReflectance = c("Constant", "Vegetation", "Clear Water", "Sand", "Lake Water"),
  EnvironmentReflectance = c("Constant", "Vegetation", "Clear Water", "Sand",
    "Lake Water"),
  TargetRadius = 0.5,
  quiet = FALSE,
  .delay = 0.5,
  .screenshot = FALSE
)

Arguments

remote_driver
RSelenium driver object, as returned by RSelenium::remoteDriver

day
The day in which the image was taken [1-31]

month
The month in which the image was taken [1-12]

SolarZenithalAngle
The solar zenith when the image was taken [degrees, between -180 and 180]

SolarAzimuthalAngle
The solar azimuth when the image was taken [degrees, between -180 and 180]

ViewZenithalAngle
The view zenith of the sensor when the image was taken [degrees, between -180 and 180]

ViewAzimuthalAngle
The view azimuth of the sensor when the image was taken [degrees, between -180 and 180]

Longitude
The longitude of the center of the image [degrees, between -180 and 180]

Latitude
The latitude of the center of the image [degrees, between -90 and 90]

Uw
Value of Uw [g/cm^-2]. Can be found in the Aeronet website: https://aeronet.gsfc.nasa.gov/.

Uo3
Value of Uo3 [cm/atm]. Can be found in the Aeronet website: https://aeronet.gsfc.nasa.gov/.

opticalDepth
Value of Atmospheric Optical Depth (AOD) in 550nm [cm/atm]. Can be found in Aeronet website: https://aeronet.gsfc.nasa.gov/.

LowerWavelength
The lowest wavelength of the spectral band for correction

UpperWavelength
The highest wavelength of the spectral band for correction

TargetAltitude
The height of the center of the image above sea level [Km]

GroundCondition
The ground condition of the image, one of: "Homogenous Ground", "Patchy Ground".

TargetReflectance
The target's reflectance type, one of: "Constant", "Vegetation", "Clear Water", "Sand", "Lake Water".

EnvironmentReflectance
The environment reflectance type of the image, one of: "Constant", "Vegetation", "Clear Water", "Sand", "Lake Water".

TargetRadius
The area of interest in the image [Km, the default is 0.5]

quiet
Should progress messages be suppressed? (Default is FALSE)

.delay
Delay between each Selenium server step (for debugging purposes)

.screenshot
Logical, if TRUE then a screenshot of the Selenium browser, named rselenium_screenshot.png, is written in the working directory at each step (for debugging purposes)

Value

Named vector of length three, with the 6S parameters xa, xb, and xc

Details

The 6S radiative transfer model relates the top of the atmosphere reflectance or radiance (TOA) with the reflectance at satellite height, and the surface reflectance, using information about the atmospheric conditions, such as aerosol optical depth (AOD) and water vapor (WV) content (Vermote et al. 1997). The atmospheric correction coefficients (Xa, Xb, Xc) returned by the function can be used to calculate bottom of the atmosphere (BOM) reflectance from top of the atmosphere (TOA) radiance using the following formulas (1) and (2) (Zhao, Tamura, and Takahashi 2001):

(1) y = Xa*TOA-Xb

(2) BOA = y / (1+Xc*y)

where 'y' is an intermediate variable, 'TOA' is the top of the atmosphere radiance, 'BOA' is the bottom of the atmosphere reflectance, and 'Xa', 'Xb', 'Xc' are the correction coefficents returned by the sixs_params function.

References

Vermote, E. F., Tanre, D., Deuze, J. L., Herman, M., & Morcette, J. J. (1997). Second simulation of the satellite signal in the solar spectrum, 6S: An overview. IEEE transactions on geoscience and remote sensing, 35(3), 675-686.

Zhao, W., Tamura, M., & Takahashi, H. (2001). Atmospheric and spectral corrections for estimating surface albedo from satellite data using 6S code. Remote sensing of Environment, 76(2), 202-212.

Examples

if (FALSE) {

library(RSelenium)

# Connect to Selenium server
remote_driver = remoteDriver(remoteServerAddr = "164.90.191.95", port = 4445L)
remote_driver$open(silent = TRUE)

# Test '6S'
xcoefficients = sixs_params(
  remote_driver = remote_driver, 
  day = 7,
  month = 7, 
  SolarZenithalAngle = 30, 
  SolarAzimuthalAngle = 117,
  ViewZenithalAngle = 25,
  ViewAzimuthalAngle = 112, 
  Longitude = 31, 
  Latitude = 34, 
  Uw = 1.52, 
  Uo3 = 0.38, 
  opticalDepth = 0.5, 
  LowerWavelength = 0.4, 
  UpperWavelength = 0.44,
  TargetAltitude = 0.342,
  GroundCondition = "Patchy Ground",
  TargetReflectance = "Vegetation",
  EnvironmentReflectance = "Vegetation", 
  TargetRadius = 0.5,
  .screenshot = TRUE
)

# Check result
xcoefficients == c(0.00627, 0.4462, 0.1665)

# Close connection
remote_driver$close()

}